requirements/RQ_LabelService.txt
django
djangorestframework-jsonapi
django-background-tasks
pymongo
confluent_kafka
django-cors-headers
happybase
pillow
numpy
base64